Make sure detecting metal rod won't work
In the previous commit message, I listed three ways to parse the video:
1. detect metal rod
2. check for total coincidence of frames
3. do that half-manually

This commit shows that the first two ways are very unlikely to work
smoothly and, unfortunately, the third option is the optimal one.

In this commit video_parser_prototype.py plots sums of squares of
differences between current frame and the background frame I saved.
It was clear from the plot (I didn't save it) that I won't be able to
detect anything that way.